Bloomagain/agenting/requirements.txt
flask
python-dotenv
langchain
langchain-ibm
langchain-community
chromadb
beautifulsoup4
requests

